c-family: Simplify attribute exclusion handling
This patch changes the handling of mutual exclusions involving the
target and target_clones attributes to use the generic attribute
exclusion lists.  Additionally, the duplicate handling for the
always_inline and noinline attribute exclusion is removed.

The only change in functionality is the choice of warning message
displayed - due to either a change in the wording for mutual exclusion
warnings, or a change in the order in which different checks occur.

gcc/c-family/ChangeLog:

	* c-attribs.cc (attr_always_inline_exclusions): New.
	(attr_target_exclusions): Ditto.
	(attr_target_clones_exclusions): Ditto.
	(c_common_attribute_table): Add new exclusion lists.
	(handle_noinline_attribute): Remove custom exclusion handling.
	(handle_always_inline_attribute): Ditto.
	(handle_target_attribute): Ditto.
	(handle_target_clones_attribute): Ditto.

gcc/testsuite/ChangeLog:

	* g++.target/i386/mvc2.C:
	* g++.target/i386/mvc3.C: