testsuite: Add testcase for already fixed PR [PR111068]
Hi!

This one unfortunately can't be bisected, it ICEd until r14-3430
inclusive, but r14-3431 removed -mavx10.1-512 support and when it
was readded in r14-5607 it doesn't ICE anymore.

I'm just committing the testcase so that it doesn't reappear.

Regtested on x86_64-linux and i686-linux, committed to trunk as obvious.

2023-12-07  Jakub Jelinek  <jakub@redhat.com>

	PR target/111068
	* gcc.target/i386/pr111068.c: New test.

	Jakub