A new copy propagation and PHI elimination pass
> > Hi,
> >
> > this is a patch that I submitted two months ago as an RFC. I added some polish
> > since.
> >
> > It is a new lightweight pass that removes redundant PHI functions and as a
> > bonus does basic copy propagation. With Jan Hubička we measured that it is able
> > to remove usually more than 5% of all PHI functions when run among early passes
> > (sometimes even 13% or more). Those are mostly PHI functions that would be
> > later optimized away but with this pass it is possible to remove them early
> > enough so that they don't get streamed when runing LTO (and also potentially
> > inlined at multiple places). It is also able to remove some redundant PHIs
> > that otherwise would still be present during RTL expansion.
> >
> > Jakub Jelínek was concerned about debug info coverage so I compiled cc1plus
> > with and without this patch. These are the sizes of .debug_info and
> > .debug_loclists
> >
> > .debug_info without patch 181694311
> > .debug_info    with patch 181692320
> > +0.0011% change
> >
> > .debug_loclists without patch 47934753
> > .debug_loclists    with patch 47934966
> > -0.0004% change
> >
> > I wanted to use dwlocstat to compare debug coverages but didn't manage to get
> > the program working on my machine sadly. Hope this suffices. Seems to me that
> > my patch doesn't have a significant impact on debug info.
> >
> > Bootstraped and tested* on x86_64-pc-linux-gnu.
> >
> > * One testcase (pr79691.c) did regress. However that is because the test is
> > dependent on a certain variable not being copy propagated. I will go into more
> > detail about this in a reply to this mail.
> >
> > Ok to commit?
>
> This is a second version of the patch.  In this version, I modified the
> pr79691.c testcase so that it works as intended with other changes from the
> patch.
>
> The pr79691.c testcase checks that we get constants from snprintf calls and
> that they simplify into a single constant.  The testcase doesn't account for
> the fact that this constant may be further copy propagated which is exactly
> what happens with this patch applied.
>
> Bootstrapped and tested on x86_64-pc-linux-gnu.
>
> Ok to commit?

This is the third version of the patch. In this version, I adressed most of
Richards remarks about the second version. Here is a summary of changes I made:

- Rename the pass from tree-ssa-sccopy.cc to gimple-ssa-sccopy.cc
- Use simple_dce_from_worklist to remove propagated statements
- Use existing replace_uses API instead of reinventing it
  - This allowed me to get rid of some now redundant cleanup code
- Encapsulate the SCC finding into a class
- Rework stmt_may_generate_copy to get rid of redundant checks
- Add check that PHI doesn't contain two non-SSA-name values to
  stmt_may_generate_copy
- Regarding alignment and value ranges in stmt_may_generate_copy: For now use
  the conservative check that Richard suggested
- Index array of vertices that SCC discovery uses by SSA name version numbers
  instead of numbering statements myself.

I didn't make any changes based on these remarks:

1 It might be nice to optimize SCCs of size 1 somehow, not sure how
  many times these appear - possibly prevent them from even entering
  the SCC discovery?

It would be nice. But the only way to do this that I see right now is to first
propagate SCCs of size 1 and then the rest. This would mean adding a new copy
propagation procedure. It wouldn't be a trivial procedure. Efficiency of the
pass relies on having SCCs topogically sorted so this procedure would have to
implement some topological sort algorithm.

This could be done. It could save allocating some vec<>s (right now, SCCs of
size 1 are represented by a vec<> with a single element). But is it worth it to
optimize the pass this way right now? If possible, I'd like to see that the
pass works and sort out any problems people encounter with it before I start
optimizing it.

2 Instead of collecting all stmts that may generate a copy at the beginning of
  the pass into a vec<>, let the SCC discovery check that statements may
  generate a copy on the fly.

This would be a big change to the pass, it would require a lot of reworking.
I'm also not sure if this would help reduce the number of allocated vec<>s that
much because I'll still want to represent SCCs by vec<>s.

Again - its possible I'll want to rework the pass in this way in the future but
I'd like to leave it as it is for now.

3 Add a comment saying that the pass is doing optimistic copy propagation

I don't think the pass works in an optimistic way. It doesn't assume that all
variables are copies of each other at any point. It instead identifies copy
statements (or PHI SCCs that act as copy statements) and then for each of these
it propagates: By that I mean if a copy statement says that _3 is a copy of _2,
then it replaces all uses of _3 by _2.

But its possible that I still misinterpret what 'optimistic' means. If
mentioning that it works in an optimistic way would help readability of the
pass, I'm happy to add that comment.

Richard, is the patch ok as it is now? Or do you think that making changes 1, 2
or 3 is necessary? Or are there any other issues which should be adressed?

Filip

-- >8 --

This patch adds the strongly-connected copy propagation (SCCOPY) pass.
It is a lightweight GIMPLE copy propagation pass that also removes some
redundant PHI statements. It handles degenerate PHIs, e.g.:

_5 = PHI <_1>;
_6 = PHI <_6, _6, _1, _1>;
_7 = PHI <16, _7>;
// Replaces occurences of _5 and _6 by _1 and _7 by 16

It also handles more complicated situations, e.g.:

_8 = PHI <_9, _10>;
_9 = PHI <_8, _10>;
_10 = PHI <_8, _9, _1>;
// Replaces occurences of _8, _9 and _10 by _1

gcc/ChangeLog:

	* Makefile.in: Added sccopy pass.
	* passes.def: Added sccopy pass before LTO streaming and before
	  RTL expanstion.
	* tree-pass.h (make_pass_sccopy): Added sccopy pass.
	* gimple-ssa-sccopy.cc: New file.

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-ssa/pr79691.c: Updated scan-tree-dump to account
	  for additional copy propagation this patch adds.
	* gcc.dg/sccopy-1.c: New test.

Signed-off-by: Filip Kastl <fkastl@suse.cz>