Fix length computation for logical shifts on H8
This fixes the length computation for logical shifts on the H8/SX.

The H8/SX has a richer set of logical shifts compared to early parts in
the H8 family.  It has special 2 byte instructions for shifts by power
of two immediate values as well as a special 4 byte shift by other
immediate values.

These were never accounted for (AFIACT) in the length computation for
shifts.  Until now that's mostly just affected branch shortening.  But
an upcoming patch uses instruction lengths to select between two
potential sequences and getting these lengths wrong will cause it to
miss optimization opportunities on the H8/SX.

A slightly different version of this patch has been tested on the H8
without regressions.

Pushed to the trunk,
Jeff
commit 4ac358c619e364ad767242409765c178da9d83e0
Author: Jeff Law <jlaw@ventanamicro.com>
Date:   Sun Dec 10 09:32:55 2023 -0700

    [committed] Fix length computation for logical shifts on H8

    This fixes the length computation for logical shifts on the H8/SX.

    The H8/SX has a richer set of logical shifts compared to early parts in the H8
    family.  It has special 2 byte instructions for shifts by power of two
    immediate values as well as a special 4 byte shift by other immediate values.

    These were never accounted for (AFIACT) in the length computation for shifts.
    Until now that's mostly just affected branch shortening.  But an upcoming patch
    uses instruction lengths to select between two potential sequences and getting
    these lengths wrong will cause it to miss optimization opportunities on the
    H8/SX.

    gcc
            * config/h8300/h8300.cc (compute_a_shift_length): Fix computation
            of logical shifts on the H8/SX.