Support uaddv and usubv on the H8
This patch adds uaddv/usubv support on the H8 port to speed up those
pesky builtin-overflow tests.  It's a variant of something I'd been
running for a while -- the major change between the old approach I'd
been using and this patch is this version does not expose the CC
register until after reload to be consistent with the rest of the H8 port.

The general approach is to first clear the GPR that's going to hold the
overflow status, perform the arithmetic operation (add/sub), then use
addx to move the overflow indicator (in the C bit) into the GPR holding
the overflow status.

That's a significant improvement over the mess of logicals that's
generated by the generic code.

Handling signed overflow is possible and something I'll probably port to
this scheme at some point.  It's a bit more complex because we can't
trivially move the bit from CCR into the right position in a GPR and
other quirks of the H8.

This has been regression tested on the H8 without problems.  Pushing to
the trunk.

Jeff
commit 7fb9454c748632d148a07c275ea1f77b290b0c2d
Author: Jeff Law <jlaw@ventanamicro.com>
Date:   Sun Dec 10 10:41:05 2023 -0700

    [committed] Support uaddv and usubv on the H8

    This patch adds uaddv/usubv support on the H8 port to speed up those pesky
    builtin-overflow tests.  It's a variant of something I'd been running for a
    while -- the major change between the old approach I'd been using and this
    patch is this version does not expose the CC register until after reload to be
    consistent with the rest of the H8 port.

    The general approach is to first clear the GPR that's going to hold the
    overflow status, perform the arithmetic operation (add/sub), then use addx to
    move the overflow indicator (in the C bit) into the GPR holding the overflow
    status.

    That's a significant improvement over the mess of logicals that's generated by
    the generic code.

    Handling signed overflow is possible and something I'll probably port to this
    scheme at some point.  It's a bit more complex because we can't trivially move
    the bit from CCR into the right position in a GPR and other quirks of the H8.

    This has been regression tested on the H8 without problems.  Pushing to the
    trunk.

    gcc/
            * config/h8300/addsub.md (uaddv<mode>4, usubv<mode>4): New expanders.
            (uaddv): New define_insn_and_split plus post-reload pattern.