Avoid losing MEM_REF offset in MEM_EXPR adjustment for stack slot sharing
When investigating PR111591 with respect to TBAA and stack slot sharing
I noticed we're eventually scrapping a [TARGET_]MEM_REF offset when
rewriting the VAR_DECL base of the MEM_EXPR to use a pointer to the
partition instead.  The following makes sure to preserve that.

Bootstrapped and tested on x86_64-unknown-linux-gnu, pushed.

	* emit-rtl.cc (set_mem_attributes_minus_bitpos): Preserve
	the offset when rewriting an exising MEM_REF base for
	stack slot sharing.