middle-end/111975 - dump -> GIMPLE FE roundtrip improvements
The following improves the manual work needed to make a -gimple dump
valid input to the GIMPLE FE.  First of all it recognizes the 'sizetype'
tree and dumps it as __SIZETYPE__, then it changes dumping vector types
without name from 'vector(n) T' to 'T [[gnu::vector_size(n')]]' which
we can parse in most relevant contexts (and that's shorter than
using __attribute__).  Third it avoids a NULL_TREE TMR_STEP when
it would be one, an optimization that's re-done when generating RTL.

Bootstrapped and tested on x86_64-unknown-linux-gnu.

OK?

Thanks,
Richard.

	PR middle-end/111975
	* tree-pretty-print.cc (dump_generic_node): Dump
	sizetype as __SIZETYPE__ with TDF_GIMPLE.
	Dump unnamed vector types as T [[gnu::vector_size(n)]] with
	TDF_GIMPLE.
	* tree-ssa-address.cc (create_mem_ref_raw): Never generate
	a NULL STEP when INDEX is specified.