c++: bad direct reference binding [PR113064]
Bootstrappde and regtested on x86_64-pc-linux-gnu, does this look OK for
trunk?

-- >8 --

When computing a direct conversion to reference type fails and yields a
bad conversion, reference_binding incorrectly commits to that conversion
rather than attempting a conversion via a temporary.  This leads to us
rejecting the first testcase because the direct bad conversion to B&& via
the && conversion operator prevents us from considering the (good)
conversion via a temporary and the & conversion operator (and similarly
for the second more elaborate testcase).  This patch fixes this by
making reference_binding not prematurely commit to such a bad direct
conversion.  We still fall back to such a bad direct conversion if using
a temporary also fails (otherwise the diagnostic for cpp0x/explicit7.C
regresses).

	PR c++/113064

gcc/cp/ChangeLog:

	* call.cc (reference_binding): Still try a conversion via a
	temporary if a direct conversion was bad.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp0x/rv-conv4.C: New test.
	* g++.dg/cpp0x/rv-conv5.C: New test.