opts: -Werror=foo always implies -Wfoo [PR106213]
Tested x86_64-pc-linux-gnu, OK for trunk?

-- 8< --

-Werror=foo implying -Wfoo wasn't working for -Wdeprecated-copy-dtor,
because it is specified as the value 2 of warn_deprecated_copy, which shows
up as CLVC_EQUAL, which is not one of the three var_types handled by
control_warning_option.  It seems to me that we can just unconditionally
handle_generated_option, and only have special argument handling for those
types.

	PR c++/106213

gcc/ChangeLog:

	* opts-common.cc (control_warning_option): Call
	handle_generated_option for all cl_var_types.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp0x/depr-copy5.C: New test.